+++ src/main/java/com/codecool/web/repository/ProfileRepo.java
@@ -3,5 +3,5 @@ package com.codecool.web.repository;
 import com.codecool.web.model.ProfileData;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProfileRepo extends CrudRepository<ProfileData, Long> {
+public interface ProfileRepo extends CrudRepository<ProfileData, Integer> {
 }
